requirements.txt
langchain
openai
faiss-cpu
streamlit
python-dotenv
python-docx
PyMuPDF

langchain-google-genai
google-generativeai
wiki
wikipedia
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
PyPDF2
python-docx
wikipedia
tiktoken